% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseslope.R
\docType{package}
\name{analyseslope}
\alias{analyseslope}
\alias{analyseslope-package}
\title{analyseslope: A package which performs Bayesian analysis and visualization for slope parameters.}
\description{
The package contains functions which extracts slope parameters from our database,
plot these slope parameters, and analyse them. We can also identify outliers.
}
\section{analyseslope functions}{

The analyseslope functions ...
}

