% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_null_deviance.R
\name{calc_null_deviance}
\alias{calc_null_deviance}
\title{Get pseudo R^2}
\usage{
calc_null_deviance(upstream_slopes, species)
}
\arguments{
\item{gradient_analysis}{Gradient analysis table derived from slope_analysis function.}

\item{null_deviance}{Null deviance value derived from calc_null_deviance function.}

\item{parameter}{Parameter to analyse.}
}
\value{
Your pseudo R^2 value.
}
\description{
Function returns null deviance for a model.
}
