% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_dbconnect.R
\name{db_connect}
\alias{db_connect}
\title{Initialises connection with database.}
\usage{
db_connect(database)
}
\arguments{
\item{database}{Database to connect to.}
}
\description{
Initialises connection with database.
}
