% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction_table.R
\name{get_prediction_table}
\alias{get_prediction_table}
\title{Predict lake colonisation}
\usage{
get_prediction_table(upstream_slopes, gradient_analysis, parameter,
  distance = TRUE, scaled.model = TRUE)
}
\arguments{
\item{upstream_slopes}{Table containing lake parameters, ids and peresence/absence of species.}

\item{gradient_analysis}{Gradient analysis tables derived from slope_analysis function.}

\item{parameter}{Parameter for which you want likelihoods.}

\item{distance}{States whether or not you want distance included in the model.}

\item{scaled.model}{Indicates whether or not gradient_analysis was scaled when you used it.}
}
\value{
Table with true presences, false presences, true absences, false absences.
}
\description{
Gives predictions of colonisation of lakes. Similar to dispersal_likelihood function, but instead of slope parameters, returns actual colonisation alongside predictions.
}
