% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pseudoR2.R
\name{get_pseudoR2}
\alias{get_pseudoR2}
\title{Get pseudo R^2}
\usage{
get_pseudoR2(gradient_analysis, null_deviance, parameter)
}
\arguments{
\item{gradient_analysis}{Gradient analysis table derived from slope_analysis function.}

\item{null_deviance}{Null deviance value derived from calc_null_deviance function.}

\item{parameter}{Parameter to analyse.}
}
\value{
Your pseudo R^2 value.
}
\description{
Function returns pseduo R^2 value for your parameter based on your model and null deviance.
}
