% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_outliers.R
\name{identify_outliers}
\alias{identify_outliers}
\title{Identifies outliers on a graph for a certain parameter.}
\usage{
identify_outliers(upstream_slopes_test, parameter_of_interest, species)
}
\arguments{
\item{upstream_slopes_test}{A data frame generated by the @extract_slope_params function.}

\item{parameter_of_interest}{The slope parameter you wish to investigate.}

\item{species}{The species to analyse}
}
\value{
A subset of @upstream_slopes_test which contains only outliers.
}
\description{
Identifies outliers on a graph for a certain parameter.
}
\examples{
identify_outliers(slope_max, upstream_slopes_test)
# Will have to zoom on graph and click outliers
}
