% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dispersal_curves.R
\name{plot_dispersal_curves}
\alias{plot_dispersal_curves}
\title{Plots dispersal curves}
\usage{
plot_dispersal_curves(your_upstream_slopes, slope_analysis_fish, species,
  slope_parameters)
}
\arguments{
\item{your_upstream_slopes}{Table generated from the extract_slope_params function.}

\item{slope_analysis_fish}{The slope analysis table for your species}

\item{species}{The species to plot}

\item{slope_parameters}{The parameters you wish to plot}
}
\description{
Creates a series of plots comparing slope parameters to presence of fish upstream.
}
