% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slope_params.R
\name{plot_slope_params}
\alias{plot_slope_params}
\title{Plots slope parameter}
\usage{
plot_slope_params(upstream_slopes_test, dimen, species, parameters_of_interest,
  main.lab = "", type = "scatter")
}
\arguments{
\item{upstream_slopes_test}{Table generated from the extract_slope_params function.}

\item{species}{The species to plot}

\item{type}{Type of plot to create. Current options are "scatter" and "boxplot".}

\item{dim}{The dimensions your plots should appear in}
}
\description{
Creates a series of plots comparing slope parameters to presence of fish upstream.
}
