% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_check.R
\name{prediction_check}
\alias{prediction_check}
\title{Check predictive accuracy}
\usage{
prediction_check(p_table, threshold)
}
\arguments{
\item{p_table}{Table with two columns, one showing likelihood of colonisation and one showing whether or not there was in fact colonisation}

\item{threshold}{Percentage threshold above which species will colonise a lake and below which they will not}
}
\value{
Table with true presences, false presences, true absences, false absences.
}
\description{
Gives predictive accuracy of a model.
}
