% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_validate.R
\name{species_validate}
\alias{species_validate}
\title{Species model validation}
\usage{
species_validate(predictions_results, fp = 2, fn = 1)
}
\arguments{
\item{predictions_results}{Data frame which lists the chance that species will colonise a lake, and whether or not they in fact did.}

\item{fp}{The cost of a false presence.}

\item{fn}{Cost of a false negative.}
}
\value{
Your pseudo R^2 value.
}
\description{
Function gives best threshold above and below of which species will or will not colonise
}
