% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slope_params.R
\name{plot_slope_params}
\alias{plot_slope_params}
\title{Creates a series of plots comparing slope parameters to presence of fish upstream.}
\usage{
plot_slope_params(upstream_slopes_test)
}
\arguments{
\item{upstream_slopes_test}{Table generated from the extract_slope_params function.

Note: this function assumes that you want to plot all the slope parameters you derived for the upstream_slopes_test table.}
}
\description{
Creates a series of plots comparing slope parameters to presence of fish upstream.
}
