% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slope_params.R
\name{plot_slope_params}
\alias{plot_slope_params}
\title{Plots slope parameter}
\usage{
plot_slope_params(upstream_slopes_test, dim, species)
}
\arguments{
\item{upstream_slopes_test}{Table generated from the extract_slope_params function.}

\item{dim}{The dimensions your plots should appear in}

\item{species}{The species to plot}
}
\description{
Creates a series of plots comparing slope parameters to presence of fish upstream.
}
