% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_dbconnect.R
\name{slope_dbconnect}
\alias{slope_dbconnect}
\title{Initialises connection with database.}
\usage{
slope_dbconnect(database, server)
}
\arguments{
\item{database}{Database to connect to.}
}
\description{
Initialises connection with database.
}
